feat(package) add setup script and configuration
Nota bene:

edx-gea is supposed to work integrated in an Open edX instance where
xblock-utils is already installed. xblock-utils is required by edx-gea
but cannot be defined as a base requirement as it is not published in
PyPI. To avoid breaking our packaging CI, we define xblock-utils as an
extra requirement labeled "edx".